requirements.txt
fastapi
uvicorn
prisma
paho-mqtt==1.6.1